bf16 fix

+++ src/metrics/metricszoo.py
@@ -25,8 +25,8 @@ class Acc1(BaseMetric):
         self.answers.append(t)
 
     def summarize(self):
-        scores = torch.cat(self.scores).cpu()
-        answers = torch.cat(self.answers).cpu().numpy()
+        scores = torch.cat(self.scores).cpu().float()
+        answers = torch.cat(self.answers).cpu().float().numpy()
 
         if scores.size(-1) > 1: # multi-class
             labels = scores.argmax(-1).numpy()
@@ -52,8 +52,8 @@ class Acc5(BaseMetric):
         self.answers.append(t)
 
     def summarize(self):
-        scores = torch.cat(self.scores).cpu().softmax(-1).numpy()
-        answers = torch.cat(self.answers).cpu().numpy()
+        scores = torch.cat(self.scores).cpu().float().softmax(-1).numpy()
+        answers = torch.cat(self.answers).cpu().float().numpy()
         num_classes = scores.shape[-1]
         return top_k_accuracy_score(answers, scores, k=5, labels=np.arange(num_classes))
 
@@ -68,8 +68,8 @@ class Auroc(BaseMetric):
         self.answers.append(t)
 
     def summarize(self):
-        scores = torch.cat(self.scores).cpu().softmax(-1).numpy()
-        answers = torch.cat(self.answers).cpu().numpy()
+        scores = torch.cat(self.scores).cpu().float().softmax(-1).numpy()
+        answers = torch.cat(self.answers).cpu().float().numpy()
         num_classes = scores.shape[-1]
         return roc_auc_score(answers, scores, average='weighted', multi_class='ovr', labels=np.arange(num_classes))
 
@@ -84,8 +84,8 @@ class Auprc(BaseMetric): # only for binary classification
         self.answers.append(t)
 
     def summarize(self):
-        scores = torch.cat(self.scores).cpu().sigmoid().numpy()
-        answers = torch.cat(self.answers).cpu().numpy()
+        scores = torch.cat(self.scores).cpu().float().sigmoid().numpy()
+        answers = torch.cat(self.answers).cpu().float().numpy()
         return average_precision_score(answers, scores, average='weighted')
 
 class Youdenj(BaseMetric):  # only for binary classification
@@ -99,8 +99,8 @@ class Youdenj(BaseMetric):  # only for binary classification
         self.answers.append(t)
 
     def summarize(self):
-        scores = torch.cat(self.scores).cpu().sigmoid().numpy()
-        answers = torch.cat(self.answers).cpu().numpy()
+        scores = torch.cat(self.scores).cpu().float().sigmoid().numpy()
+        answers = torch.cat(self.answers).cpu().float().numpy()
         fpr, tpr, thresholds = roc_curve(answers, scores)
         return thresholds[np.argmax(tpr - fpr)]
 
@@ -116,8 +116,8 @@ class F1(BaseMetric):
         self.answers.append(t)
 
     def summarize(self):
-        scores = torch.cat(self.scores).cpu()
-        answers = torch.cat(self.answers).cpu().numpy()
+        scores = torch.cat(self.scores).cpu().float()
+        answers = torch.cat(self.answers).cpu().float().numpy()
 
         if scores.size(-1) > 1: # multi-class
             labels = scores.argmax(-1).numpy()
@@ -143,8 +143,8 @@ class Precision(BaseMetric):
         self.answers.append(t)
 
     def summarize(self):
-        scores = torch.cat(self.scores).cpu()
-        answers = torch.cat(self.answers).cpu().numpy()
+        scores = torch.cat(self.scores).cpu().float()
+        answers = torch.cat(self.answers).cpu().float().numpy()
 
         if scores.size(-1) > 1: # multi-class
             labels = scores.argmax(-1).numpy()
@@ -170,8 +170,8 @@ class Recall(BaseMetric):
         self.answers.append(t)
 
     def summarize(self):
-        scores = torch.cat(self.scores).cpu()
-        answers = torch.cat(self.answers).cpu().numpy()
+        scores = torch.cat(self.scores).cpu().float()
+        answers = torch.cat(self.answers).cpu().float().numpy()
 
         if scores.size(-1) > 1: # multi-class
             labels = scores.argmax(-1).numpy()
@@ -197,8 +197,8 @@ class Seqacc(BaseMetric):
         self.answers.append(t.view(-1))
 
     def summarize(self):
-        labels = torch.cat(self.scores).cpu().argmax(-1).numpy()
-        answers = torch.cat(self.answers).cpu().numpy()
+        labels = torch.cat(self.scores).cpu().float().argmax(-1).numpy()
+        answers = torch.cat(self.answers).cpu().float().numpy()
 
         # ignore special tokens
         labels = labels[answers != -1]
@@ -217,8 +217,8 @@ class Mse(BaseMetric):
         self.answers.append(t)
 
     def summarize(self):
-        scores = torch.cat(self.scores).cpu().numpy()
-        answers = torch.cat(self.answers).cpu().numpy()
+        scores = torch.cat(self.scores).cpu().float().numpy()
+        answers = torch.cat(self.answers).cpu().float().numpy()
         return mean_squared_error(answers, scores)
 
 class Rmse(Mse):
@@ -226,8 +226,8 @@ class Rmse(Mse):
         super(Rmse, self).__init__()
 
     def summarize(self):
-        scores = torch.cat(self.scores).cpu().numpy()
-        answers = torch.cat(self.answers).cpu().numpy()
+        scores = torch.cat(self.scores).cpu().float().numpy()
+        answers = torch.cat(self.answers).cpu().float().numpy()
         return mean_squared_error(answers, scores, squared=False)
 
 class Mae(BaseMetric):
@@ -241,8 +241,8 @@ class Mae(BaseMetric):
         self.answers.append(t)
 
     def summarize(self):
-        scores = torch.cat(self.scores).cpu().numpy()
-        answers = torch.cat(self.answers).cpu().numpy()
+        scores = torch.cat(self.scores).cpu().float().numpy()
+        answers = torch.cat(self.answers).cpu().float().numpy()
         return mean_absolute_error(answers, scores)
 
 class Mape(BaseMetric):
@@ -256,8 +256,8 @@ class Mape(BaseMetric):
         self.answers.append(t)
 
     def summarize(self):
-        scores = torch.cat(self.scores).cpu().numpy()
-        answers = torch.cat(self.answers).cpu().numpy()
+        scores = torch.cat(self.scores).cpu().float().numpy()
+        answers = torch.cat(self.answers).cpu().float().numpy()
         return mean_absolute_percentage_error(answers, scores)
 
 class R2(BaseMetric):
@@ -271,8 +271,8 @@ class R2(BaseMetric):
         self.answers.append(t)
 
     def summarize(self, *args):
-        scores = torch.cat(self.scores).cpu().numpy()
-        answers = torch.cat(self.answers).cpu().numpy()
+        scores = torch.cat(self.scores).cpu().float().numpy()
+        answers = torch.cat(self.answers).cpu().float().numpy()
         return r2_score(answers, scores)
 
 class D2(BaseMetric):
@@ -286,8 +286,8 @@ class D2(BaseMetric):
         self.answers.append(t)
 
     def summarize(self, *args):
-        scores = torch.cat(self.scores).cpu().numpy()
-        answers = torch.cat(self.answers).cpu().numpy()
+        scores = torch.cat(self.scores).cpu().float().numpy()
+        answers = torch.cat(self.answers).cpu().float().numpy()
         return d2_pinball_score(answers, scores)
 
 
@@ -303,8 +303,8 @@ class Mlacc(BaseMetric):
         self.answers.append(t)
 
     def summarize(self):
-        scores = torch.cat(self.scores).cpu().sigmoid().numpy()
-        answers = torch.cat(self.answers).cpu().numpy()
+        scores = torch.cat(self.scores).cpu().float().sigmoid().numpy()
+        answers = torch.cat(self.answers).cpu().float().numpy()
         preds = (scores >= 0.5).astype(float)
         # per-label accuracy, then average
         return np.mean((preds == answers).mean(axis=0))
@@ -322,8 +322,8 @@ class Mlauroc(BaseMetric):
         self.answers.append(t)
 
     def summarize(self):
-        scores = torch.cat(self.scores).cpu().sigmoid().numpy()
-        answers = torch.cat(self.answers).cpu().numpy()
+        scores = torch.cat(self.scores).cpu().float().sigmoid().numpy()
+        answers = torch.cat(self.answers).cpu().float().numpy()
         # per-label AUROC, skip labels with single class
         aurocs = []
         for i in range(answers.shape[1]):
